% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvs_get_glob.R
\name{dvs_get_glob}
\alias{dvs_get_glob}
\title{Retrieve data files from \code{dvs} storage}
\usage{
dvs_get_glob(path = NULL, glob = NULL, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[fs:dir_ls]{fs::dir_ls}}
  \describe{
    \item{\code{path}}{A character vector of one or more paths.}
    \item{\code{all}}{If \code{TRUE} hidden files are also returned.}
    \item{\code{recurse}}{If \code{TRUE} recurse fully, if a positive number the number of levels
to recurse.}
    \item{\code{type}}{File type(s) to return, one or more of "any", "file", "directory",
"symlink", "FIFO", "socket", "character_device" or "block_device".}
    \item{\code{glob}}{A wildcard aka globbing pattern (e.g. \verb{*.csv}) passed on to \code{\link[=grep]{grep()}} to filter paths.}
    \item{\code{regexp}}{A regular expression (e.g. \verb{[.]csv$}) passed on to \code{\link[=grep]{grep()}} to filter paths.}
    \item{\code{invert}}{If \code{TRUE} return files which do \emph{not} match}
    \item{\code{fail}}{Should the call fail (the default) or warn if a file cannot be
accessed.}
    \item{\code{recursive}}{(Deprecated) If \code{TRUE} recurse fully.}
  }}

\item{files}{character vector containing file paths}
}
\description{
Retrieve data files from \code{dvs} storage
}
