% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvs.R
\name{dvs_materialize}
\alias{dvs_materialize}
\title{Materialize Files from Cache}
\usage{
dvs_materialize()
}
\value{
A list with materialize summary (materialized, up_to_date, failed, results).
}
\description{
Copy cached objects to their working tree locations based on the manifest.
}
\examples{
\dontrun{
dvs_materialize()
}
}
