% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvs-wrappers.R
\name{num_cpus}
\alias{num_cpus}
\title{Detect cores for parallelism}
\source{
Generated by miniextendr from Rust fn \code{num_cpus}
}
\usage{
num_cpus()
}
\description{
Returns the number of logical cores currently available in the system.
}
\note{
Unlike \code{parallel::detectCores()}, \code{num_cpus()} respects cgroups.
}
