% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvs-wrappers.R
\name{dvs_test_2_5_gib_bytes}
\alias{dvs_test_2_5_gib_bytes}
\title{test support}
\source{
Generated by miniextendr from Rust fn \code{dvs_test_2_5_gib_bytes}
}
\usage{
dvs_test_2_5_gib_bytes()
}
\description{
test support
}
