% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvs-package.R
\docType{package}
\name{dvs-package}
\alias{dvs}
\alias{dvs-package}
\title{dvs: Data Version Control System by A2-AI}
\description{
Provides version control for large binary and text files that are stored either locally or remotely. Uses content-addressable storage with blake3 hashing to efficiently deduplicate files while tracking metadata in Git-friendly .dvs files.
}
\author{
\strong{Maintainer}: Jenna Johnson \email{jenna@a2-ai.com}

Authors:
\itemize{
  \item Devin Pastoor \email{devin@a2-ai.com}
  \item Mossa Merhi Reimert \email{mossa@a2-ai.com}
}

Other contributors:
\itemize{
  \item A2-Ai [copyright holder, funder]
}

}
\keyword{internal}
