% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvs-commands.R
\name{dvs_init}
\alias{dvs_init}
\title{Initialize a DVS repository in the given directory}
\usage{
dvs_init(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=dvs_init_impl]{dvs_init_impl}}
  \describe{
    \item{\code{storage_path}}{Path to the storage directory where file contents are kept.}
    \item{\code{root_dir}}{Repository root directory. Defaults to the current working directory.}
    \item{\code{group}}{Unix group name to set on stored files for shared access.}
    \item{\code{metadata_folder_name}}{Name of the metadata folder. Defaults to \code{.dvs}.}
    \item{\code{no_compression}}{If \code{TRUE}, disable compression for stored files.}
  }}
}
\description{
Initialize a DVS repository in the given directory
}
