% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvs-commands.R
\name{dvs_status}
\alias{dvs_status}
\title{Report the sync status of DVS-managed files}
\usage{
dvs_status(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=dvs_status_impl]{dvs_status_impl}}
  \describe{
    \item{\code{current}}{If \code{TRUE}, include files whose local copy matches storage.}
    \item{\code{absent}}{If \code{TRUE}, include files that exist in metadata but not locally.}
    \item{\code{unsynced}}{If \code{TRUE}, include files whose local copy differs from storage.}
  }}
}
\description{
Report the sync status of DVS-managed files
}
