% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvs-wrappers.R
\name{format_byte_size}
\alias{format_byte_size}
\title{format byte size}
\source{
Generated by miniextendr from Rust fn \code{format_byte_size}
}
\usage{
format_byte_size(size_bytes)
}
\arguments{
\item{size_bytes}{(no documentation available)}
}
\description{
format byte size
}
