% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvs-wrappers.R
\docType{data}
\name{ProgressBarCallback}
\alias{ProgressBarCallback}
\alias{ProgressBarCallback$new}
\alias{$.ProgressBarCallback}
\alias{[[.ProgressBarCallback}
\title{ProgressBarCallback  Class}
\format{
An object of class \code{environment} of length 1.
}
\source{
Generated by miniextendr from Rust type \code{ProgressBarCallback}

Generated by miniextendr from \code{ProgressBarCallback::new}
}
\usage{
ProgressBarCallback

\method{$}{ProgressBarCallback}(self, name)

\method{[[}{ProgressBarCallback}(self, name)
}
\arguments{
\item{self}{The object instance.}

\item{name}{Method name for dispatch.}
}
\description{
Create a progress callback handle (signals that progress should be shown).
}
\keyword{internal}
